EE-441: Merge define and call contracts.
