+++ execution-engine/Cargo.toml
@@ -17,7 +17,6 @@ members = [
     "test-contracts/local-state",
     "test-contracts/known-urefs",
     "test-contracts/ee-441-rng-state",
-    "test-contracts/ee-441-rng-state-call",
     "test-contracts/get-caller",
     "test-contracts/get-caller-subcall",
     "wasm-prep",
