+++ execution-engine/scripts/run-contract-tests.sh
@@ -13,7 +13,6 @@ CONTRACTS=(
     "local-state"
     "known-urefs"
     "ee-441-rng-state"
-    "ee-441-rng-state-call"
     "get-caller"
     "get-caller-subcall"
 )
