--- execution-engine/test-contracts/ee-441-rng-state-call/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "ee-441-rng-state-call"
-version = "0.1.0"
-authors = ["Henry Till <henrytill@gmail.com>"]
-edition = "2018"
-
-[lib]
-name = "ee_441_rng_state_call"
-crate-type = ["cdylib"]
-
-[features]
-default = []
-std = ["cl_std/std"]
-
-[dependencies]
-cl_std = { path = "../../common", package = "casperlabs-contract-ffi" }
