+++ execution-engine/test-contracts/ee-441-rng-state/src/lib.rs
@@ -3,7 +3,6 @@
 
 #[macro_use]
 extern crate alloc;
-
 extern crate cl_std;
 
 use alloc::collections::btree_map::BTreeMap;
@@ -11,6 +10,10 @@ use alloc::string::String;
 
 use cl_std::contract_api;
 use cl_std::contract_api::pointers::ContractPointer;
+use cl_std::contract_api::{add_uref, get_arg, new_uref};
+use cl_std::key::Key;
+use cl_std::uref::URef;
+use cl_std::value::U512;
 
 #[no_mangle]
 pub extern "C" fn do_nothing() {
@@ -29,9 +32,29 @@ pub extern "C" fn do_something() {
 
 #[no_mangle]
 pub extern "C" fn call() {
-    // Export two functions
-    let _contract_pointer: ContractPointer =
-        contract_api::store_function("do_nothing", BTreeMap::new());
-    let _contract_pointer: ContractPointer =
+    let flag: String = get_arg(0);
+    let do_nothing: ContractPointer = contract_api::store_function("do_nothing", BTreeMap::new());
+    let do_something: ContractPointer =
         contract_api::store_function("do_something", BTreeMap::new());
+    if flag == "pass1" {
+        // Two calls should forward the internal RNG. This pass is a baseline.
+        let uref1: URef = new_uref(U512::from(0)).into();
+        let uref2: URef = new_uref(U512::from(1)).into();
+        add_uref("uref1", &Key::URef(uref1));
+        add_uref("uref2", &Key::URef(uref2));
+    } else if flag == "pass2" {
+        let uref1: URef = new_uref(U512::from(0)).into();
+        add_uref("uref1", &Key::URef(uref1));
+        // do_nothing doesn't do anything. It SHOULD not forward the internal RNG.
+        let result: String = contract_api::call_contract(do_nothing.clone(), &(), &vec![]);
+        assert_eq!(result, "Hello, world!");
+        let uref2: URef = new_uref(U512::from(1)).into();
+        add_uref("uref2", &Key::URef(uref2));
+    } else if flag == "pass3" {
+        let uref1: URef = new_uref(U512::from(0)).into();
+        add_uref("uref1", &Key::URef(uref1));
+        // do_something returns a new uref, and it should forward the internal RNG.
+        let uref2: URef = contract_api::call_contract(do_something.clone(), &(), &vec![]);
+        add_uref("uref2", &Key::URef(uref2));
+    }
 }
